# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Generated API",
    "description": "REST API generated by Backend Creator (bcm)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/units": {
      "get": {
        "tags": [
          "Unit"
        ],
        "summary": "List all units",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for @bcm.searchable fields"
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relation names to include"
          }
        ],
        "responses": {
          "200": {
            "description": "List of units",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UnitResponse"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Unit"
        ],
        "summary": "Create a new unit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Unit created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"ADMIN\""
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/units/{id}": {
      "get": {
        "tags": [
          "Unit"
        ],
        "summary": "Get unit by key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unit details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": [
          "Unit"
        ],
        "summary": "Update unit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unit updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Unit"
        ],
        "summary": "Partially update unit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unit updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Unit"
        ],
        "summary": "Delete unit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/sellUnitRequests": {
      "get": {
        "tags": [
          "SellUnitRequest"
        ],
        "summary": "List all sellUnitRequests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for @bcm.searchable fields"
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relation names to include"
          }
        ],
        "responses": {
          "200": {
            "description": "List of sellUnitRequests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SellUnitRequestResponse"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SellUnitRequest"
        ],
        "summary": "Create a new sellUnitRequest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellUnitRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SellUnitRequest created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellUnitRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/sellUnitRequests/{id}": {
      "get": {
        "tags": [
          "SellUnitRequest"
        ],
        "summary": "Get sellUnitRequest by key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SellUnitRequest details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellUnitRequestDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": [
          "SellUnitRequest"
        ],
        "summary": "Update sellUnitRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellUnitRequestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SellUnitRequest updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellUnitRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "SellUnitRequest"
        ],
        "summary": "Partially update sellUnitRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellUnitRequestPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SellUnitRequest updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellUnitRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "SellUnitRequest"
        ],
        "summary": "Delete sellUnitRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Unit_LocationInput": {
        "type": "object",
        "properties": {
          "create": {
            "type": "object",
            "properties": {
              "address": {
                "type": "string"
              },
              "latitude": {
                "type": "number",
                "format": "float",
                "nullable": true
              },
              "longitude": {
                "type": "number",
                "format": "float",
                "nullable": true
              },
              "deletedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            },
            "required": [
              "address"
            ],
            "description": "Create a related record"
          },
          "connect": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "description": "Connect to an existing record by unique selector"
          }
        },
        "description": "Nested input: provide either 'create' or 'connect'",
        "anyOf": [
          {
            "required": [
              "create"
            ]
          },
          {
            "required": [
              "connect"
            ]
          }
        ]
      },
      "UnitResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availability": {
            "$ref": "#/components/schemas/UnitAvailability"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "locationId": {
            "type": "string"
          },
          "acceptedSellRequestId": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UnitDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UnitResponse"
          }
        }
      },
      "UnitCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availability": {
            "$ref": "#/components/schemas/UnitAvailability"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptedSellRequestId": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/Unit_LocationInput"
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "type",
          "location"
        ]
      },
      "UnitUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availability": {
            "$ref": "#/components/schemas/UnitAvailability"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptedSellRequestId": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/Unit_LocationInput"
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "type",
          "location"
        ]
      },
      "UnitPatch": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "availability": {
            "$ref": "#/components/schemas/UnitAvailability"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptedSellRequestId": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/Unit_LocationInput"
          }
        }
      },
...
      "SellUnitRequestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "address": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SellUnitRequestDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SellUnitRequestResponse"
          }
        }
      },
      "SellUnitRequestCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "address": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SellUnitRequest_UserInput"
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "type",
          "address",
          "user"
        ]
      },
      "SellUnitRequestUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "address": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SellUnitRequest_UserInput"
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "type",
          "address",
          "user"
        ]
      },
      "SellUnitRequestPatch": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/UnitType"
          },
          "address": {
            "type": "string"
          },
          "locationId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "videoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SellUnitRequest_UserInput"
          }
        }
      },
...
      "RequestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "REJECTED",
          "CANCELLED",
          "AVAILABLE"
        ]
      },
      "UnitType": {
        "type": "string",
        "enum": [
          "RESIDENTIAL",
          "COMMERCIAL"
        ]
      },
      "UnitAvailability": {
        "type": "string",
        "enum": [
          "AVAILABLE",
          "RESERVED",
          "SOLD",
          "UNAVAILABLE"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}